+++ jannovar-cli/src/main/java/de/charite/compbio/jannovar/cmd/hgvs_to_vcf/ProjectTranscriptToChromosome.java
@@ -113,6 +113,11 @@ public class ProjectTranscriptToChromosome extends JannovarAnnotationCommand {
 		} catch (FileNotFoundException e) {
 			throw new UncheckedJannovarException("Could not load FASTA index", e);
 		}
+		if (this.fasta.getSequenceDictionary() == null) {
+			throw new UncheckedJannovarException(
+					"FASTA sequence dictionary empty, you have a REFERENCE.dict file (create with Picard "
+							+ "or samtools dict, version >=1.3)");
+		}
 
 		this.translator = new NucleotideChangeToGenomeVariantTranslator(jannovarData, fasta);
 	}
@@ -175,10 +180,33 @@ public class ProjectTranscriptToChromosome extends JannovarAnnotationCommand {
 		}
 	}
 
+	/** Map contig name (from genome variant) to contig name in FASTA */
+	private String mapContigToFasta(String contigName) {
+		// Map genome variant's contig to unique ID
+		Integer contigID = jannovarData.getRefDict().getContigNameToID().get(contigName);
+		if (contigID == null)
+			throw new UncheckedJannovarException("Unknown contig name " + contigName);
+		// Try to find matching contig in fasta
+		String nameInFasta = null;
+		for (SAMSequenceRecord record : fasta.getSequenceDictionary().getSequences()) {
+			if (jannovarData.getRefDict().getContigNameToID().containsKey(record.getSequenceName())) {
+				nameInFasta = record.getSequenceName();
+				break;
+			}
+		}
+		if (nameInFasta == null)
+			throw new UncheckedJannovarException("Could not find corresponding contig in FASTA for " + contigName);
+
+		return nameInFasta;
+	}
+
 	private void writeVariant(VariantContextWriter writer, GenomeVariant genomeVar) {
+		String nameInFasta = mapContigToFasta(genomeVar.getChrName());
 		List<Allele> alleles = new ArrayList<Allele>();
+		int shift = 0;
 		if (genomeVar.getRef().isEmpty() || genomeVar.getAlt().isEmpty()) {
-			String left = fasta.getSubsequenceAt(genomeVar.getChrName(), genomeVar.getPos(), genomeVar.getPos() + 1)
+			shift = -1;
+			String left = fasta.getSubsequenceAt(nameInFasta, genomeVar.getPos(), genomeVar.getPos())
 					.getBaseString();
 			alleles.add(Allele.create(left + genomeVar.getRef(), true));
 			alleles.add(Allele.create(left + genomeVar.getAlt(), false));
@@ -188,8 +216,8 @@ public class ProjectTranscriptToChromosome extends JannovarAnnotationCommand {
 		}
 
 		VariantContextBuilder builder = new VariantContextBuilder();
-		builder.chr(genomeVar.getChrName()).start(genomeVar.getPos() + 1)
-				.computeEndFromAlleles(alleles, genomeVar.getPos() + 1).alleles(alleles);
+		builder.chr(genomeVar.getChrName()).start(genomeVar.getPos() + shift + 1)
+				.computeEndFromAlleles(alleles, genomeVar.getPos() + shift + 1).alleles(alleles);
 
 		writer.add(builder.make());
 	}