+++ jannovar-htsjdk/src/main/java/de/charite/compbio/jannovar/htsjdk/GenomeRegionSequenceExtractor.java
@@ -1,7 +1,10 @@
 package de.charite.compbio.jannovar.htsjdk;
 
+import htsjdk.samtools.SAMSequenceRecord;
 import htsjdk.samtools.reference.IndexedFastaSequenceFile;
 import htsjdk.samtools.reference.ReferenceSequence;
+import de.charite.compbio.jannovar.UncheckedJannovarException;
+import de.charite.compbio.jannovar.data.JannovarData;
 import de.charite.compbio.jannovar.reference.GenomeInterval;
 import de.charite.compbio.jannovar.reference.Strand;
 
@@ -12,12 +15,20 @@ import de.charite.compbio.jannovar.reference.Strand;
  */
 public class GenomeRegionSequenceExtractor {
 
+	/** Jannovar database for mapping between canonical name and name in FASTA */
+	final JannovarData jannovarData;
 	/** object to load sequences from */
 	IndexedFastaSequenceFile indexedFile;
 
-	public GenomeRegionSequenceExtractor(IndexedFastaSequenceFile indexedFile) {
+	public GenomeRegionSequenceExtractor(JannovarData jannovarData, IndexedFastaSequenceFile indexedFile) {
 		super();
+		this.jannovarData = jannovarData;
 		this.indexedFile = indexedFile;
+		if (this.indexedFile.getSequenceDictionary() == null) {
+			throw new UncheckedJannovarException(
+					"FASTA file has no sequence dictionary. Are you missing the REFERENCE.dict file? "
+							+ "Hint: create with samtools dict (version >=1.2) or Picard.");
+		}
 	}
 
 	/**
@@ -30,8 +41,29 @@ public class GenomeRegionSequenceExtractor {
 	public String load(GenomeInterval region) {
 		region = region.withStrand(Strand.FWD);
 		String contigName = region.getRefDict().getContigIDToName().get(region.getChr());
+		contigName = mapContigToFasta(contigName);
 		ReferenceSequence seq = indexedFile.getSubsequenceAt(contigName, region.getBeginPos() + 1, region.getEndPos());
 		return new String(seq.getBases());
+	}
+
+	/** Map contig name (from genome variant) to contig name in FASTA */
+	private String mapContigToFasta(String contigName) {
+		// Map genome variant's contig to unique ID
+		Integer contigID = jannovarData.getRefDict().getContigNameToID().get(contigName);
+		if (contigID == null)
+			throw new UncheckedJannovarException("Unknown contig name " + contigName);
+		// Try to find matching contig in fasta
+		String nameInFasta = null;
+		for (SAMSequenceRecord record : indexedFile.getSequenceDictionary().getSequences()) {
+			if (jannovarData.getRefDict().getContigNameToID().containsKey(record.getSequenceName())) {
+				nameInFasta = record.getSequenceName();
+				break;
+			}
+		}
+		if (nameInFasta == null)
+			throw new UncheckedJannovarException("Could not find corresponding contig in FASTA for " + contigName);
+
+		return nameInFasta;
 	}
 
 }