+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 ### jannovar-cli
 
+* Fixing bug in HGVS to VCF
+* Better handling missing `.dict` file for HGVS to VCF translation
 * Adding `--annotate-as-singleton-pedigree` parameter for annotation of singleton pedigrees without pedigree file (single individual is assumed to be affected)
 * More friendly user message in case of unsorted files on inheritance mode annotation
 * Interpretation of filters in compatible inheritance mode annotation